requirements.txt
# requirements.txt
requests
paramiko
python-decouple==3.1
click
rich
pytest
